backend/requirements.txt
# LiveKit core SDK
livekit-api

# Agents and plugins
livekit-agents
livekit-plugins-google
livekit-plugins-noise-cancellation

# Flask + async support
flask[async]
flask-cors

# Env and utils
python-dotenv
requests
uuid

# Timezone
tzdata
